refactor: 댓글방 관련 로직 댓글 도메인으로 이동
