+++ backend/src/main/java/com/zzang/chongdae/offering/service/OfferingService.java
@@ -2,7 +2,6 @@ package com.zzang.chongdae.offering.service;
 
 import com.zzang.chongdae.global.exception.MarketException;
 import com.zzang.chongdae.member.repository.entity.MemberEntity;
-import com.zzang.chongdae.offering.domain.CommentRoomStatus;
 import com.zzang.chongdae.offering.domain.OfferingFilter;
 import com.zzang.chongdae.offering.domain.OfferingMeeting;
 import com.zzang.chongdae.offering.domain.OfferingPrice;
@@ -10,8 +9,6 @@ import com.zzang.chongdae.offering.domain.OfferingStatus;
 import com.zzang.chongdae.offering.exception.OfferingErrorCode;
 import com.zzang.chongdae.offering.repository.OfferingRepository;
 import com.zzang.chongdae.offering.repository.entity.OfferingEntity;
-import com.zzang.chongdae.offering.service.dto.CommentRoomInfoResponse;
-import com.zzang.chongdae.offering.service.dto.CommentRoomStatusResponse;
 import com.zzang.chongdae.offering.service.dto.OfferingAllResponse;
 import com.zzang.chongdae.offering.service.dto.OfferingAllResponseItem;
 import com.zzang.chongdae.offering.service.dto.OfferingDetailResponse;
@@ -118,28 +115,6 @@ public class OfferingService {
         offeringMemberRepository.save(offeringMember);
 
         return savedOffering.getId();
-    }
-
-    public CommentRoomInfoResponse getCommentRoomInfo(Long offeringId, MemberEntity member) {
-        // TODO : 로그인 사용자가 참여자인지 확인
-        OfferingEntity offering = offeringRepository.findById(offeringId)
-                .orElseThrow(() -> new MarketException(OfferingErrorCode.NOT_FOUND));
-        if (offering.isStatusGrouping() && offering.toOfferingStatus().isAutoConfirmed()) {
-            offering.moveStatus();
-        }
-        return new CommentRoomInfoResponse(offering, member);
-    }
-
-    @Transactional
-    public CommentRoomStatusResponse updateCommentRoomStatus(Long offeringId, MemberEntity member) {
-        // TODO: loginMember 가 총대 권한을 가지고 있는지 확인
-        OfferingEntity offering = offeringRepository.findById(offeringId)
-                .orElseThrow(() -> new MarketException(OfferingErrorCode.NOT_FOUND));
-        CommentRoomStatus updatedStatus = offering.moveStatus();
-        if (updatedStatus.equals(CommentRoomStatus.BUYING)) {
-            offering.manuallyConfirm();
-        }
-        return new CommentRoomStatusResponse(updatedStatus);
     }
 
     public OfferingProductImageResponse uploadProductImage(MultipartFile image) {
